Refactor Employee model and update respective dependencies
The Employee model was heavily refactored. The "username" and "password" fields were removed while the new "Contract" field was introduced. Also, the creation process of an Employee was modified - now an instance of Employee is not generated via static method but by using a direct constructor.
Related changes were reflected in dependent classes and interfaces. The IEmployeeRepository interface and its usage in the CreateEmployee handler class were changed to reflect the removal of the "username" field. The respective adjustments were made in the EmployeesController file as well to accommodate the updated CreateEmployee handler.
A new EmployeeRepository file was created as a placeholder for the forthcoming Employee object operations.
The changes aim to simplify the Employee object model without losing its primary semantics and to prepare a structure for further improvements over the new "Contract" model.