Refactor EmployeeRepository to use IMongoStore
The IEmployeeRepository interface has been updated, now the GetByEmail method returns a nullable Employee type to account for the possibility of an employee not found by email.
A new IMongoStore interface has been created to abstract the interactions with MongoDB. The EmployeeRepository has been updated to utilize this interface to perform its tasks, making it more loosely coupled and easier to test.
The MongoStore class has also been updated to implement the new IMongoStore interface. Here, the FindAll method now returns an IEnumerable instead of IAsyncCursor which is more generally useful. An InsertMany method was also provided to allow for insertion of multiple documents at once.