Update testing configurations and enhance Employee management
Enhancements include modifications to the TestConfiguration.cs, allowing the use of both InMemoryDatabase & MongoDb for testings as well as the addition of an IEmployeeRepository in Seeder.cs, permitting the addition of employees during data seeding.

Moreover, the RickAndMorty test suite now uses the new Employees test data and the TestFixtures now have tear-down methods for releasing resources after tests.

The Employee Controller and ViewModel are also updated to accommodate employee management operations.

A connectionString has been added to the appsettings.json for database connectivity.

Lastly, the Test.Configuration project now includes a reference to Mongo2Go to facilitate MongoDb related operations.

These changes allow us to improve the accuracy of tests and simplify the process of managing employee data in the application.