Add contract update test and refactor contract update logic
The EmployeeTests.cs has been updated to include a new test for employee contract updates. This ensures the functionality to modify an employee contract is working as expected. Moreover, the UpdateContract handler in UpdateContract.cs and the UpdateContract action in EmployeesController.cs have been refactored to return Contract objects instead of Employee objects. A system.Net.Http namespace reference has also been added to EmployeeTests.cs, possibly for handling HTTP related functions in test cases. A new Contract to EmployeeContractViewModel mapping has been introduced in EmployeeMappings.cs. This is done to cater for the aforementioned changes in the UpdateContract logic.